requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
python-dateutil>=2.8.0
regex>=2021.0.0

# Development dependencies (for testing and documentation)
# To install: pip install -r requirements-dev.txt
